[ToDatapack Phase] `goto` (untested)
As the commit title says. this is very untested. A few complex hand-made examples work, but they are not put into tests yet. In fact, due to how volatile this feature is in terms of output datapacks, the next thing I need to do might be support for in-Minecraft unit testing.
Nevertheless, the changes here are as follows:
* Added the implementation of the `GotoLabelerWalker` as all of goto depends on that data.
* Added support for goto labels and the non-switch goto keyword.
* Most of this implementation is added to `HandleIfElseGroup()`, as that is where most of the complexity comes in with how to handle the fact that minecraft functions just *keep going* and goto wants the *opposite*. For more details on this, see the `GotoLabelerWalker.AfterScopeRequiresFlag[Consume]` methods.
* Decided for convenience that `BlockOrStatement` is stupid and everything should be blocks if they make sense to be blocks.

I might at some point need to put my notebook's notes into documentation here to really clearly show what's happening why here.