`UnreachableCodeException` exception for unreachable code
Don't ask

(Actually, do. As you know, in c#, code of the form
```csharp
if (cond) { return 3; }
else { return 4; }
return 5;
```
warns you that `return 5` will never be reached. However, if you do
```csharp
bool temp = cond;
if (temp) { return 3; }
if (!temp) {return 4; }
return 5;
```
c# actually believes the `return 5` is reachable. As one of the goals is to remove all elses, this is a clear problem, so as a solution I introduce an exception the compiler considers a noop.)